==> 03_Pivoting.ipynb <==
# default_exp pivoting

#hide
from nbdev.showdoc import *

#hide
# stellt sicher, dass beim verändern der core library diese wieder neu geladen wird
%load_ext autoreload
%autoreload 2

# pivoting

In order to create uniform datasets for the different primary financial statements (Income Statement, CaschFlow, BalanceSheet, ..) it is easier if the data is converted from its vertical for to a horizontal form. This means that we pivot the values based on the tag.

Therfore, for every primary financial statement a separate dataset is created.

Statement types

- IS: IncomeStatement
- CF: CashFlow
- BS: BalanceSheet
- CI: Comprehensive Income
- EQ: Equity
- CP: CoverPage
- UN: Unclassifiable Statement

## Basic Settings

# imports
from bfh_mt_hs2020_sec_data.core import get_spark_session # initialze spark
from pathlib import Path
from typing import List, Tuple, Union, Set
from pyspark.sql.dataframe import DataFrame
from pyspark.sql.functions import *

import pandas as pd

import shutil          # provides high level file operations
import time            # used to measure execution time
import os
import sys

all_filtered_folder        = "D:/data_mt/02_filtered_parquet"               # source folder with the prepared parquet file
all_pivot_selected_folder  = "D:/data_mt/03_1_selected_pivot_cols_parquet"  # target folder which will contain only the needed subset of columns
all_pivoted_folder         = "D:/data_mt/03_2_split_pivot_parquet"          # target folder for the different pivoted and separated datasets

# init Spark
spark = get_spark_session() # Session anlegen
spark # display the moste important information of the session

## 01_Load Data

**only execute if necessary**

Loads the and filtered data with all columns

df_all = spark.read.parquet(all_filtered_folder).cache()

# Expected Amount of Data
# Entries:  15_650_848
# duration:  66 sec

# load all data into memory
start = time.time()
print("Entries: ", "{:_}".format(df_all.count())) # loading all dataset into memory
duration = time.time() - start
print("duration: ", duration)

## 02_Select

**only execute if necessary**

Creates a new dataset containing only  the columns that are needed during the next steps.

from pyspark.sql.functions import year
df_all_selected = df_all.select(["stmt","cik","ticker", "adsh","period","form","filed","tag","value","report", "line", "fp", "uom", "qtrs","negating",
                                year(df_all.period).alias('period_year')]).cache()

df_all_selected = df_all_selected.where("period_year >= 2012")
print(df_all_selected.count())

shutil.rmtree(all_pivot_selected_folder,  ignore_errors=True)
df_all_selected.write.parquet(all_pivot_selected_folder)

## 03_Pivoting

df_all_selected = spark.read.parquet(all_pivot_selected_folder).cache()

# Expected Amount of Data
# Entries:  13_713_772  # 2012-2020 q3
# duration:  23 sec

# load all data into memory
start = time.time()
print("Entries: ", "{:_}".format(df_all_selected.count())) # loading all dataset into memory
duration = time.time() - start
print("duration: ", duration)

# if a value is null in the original data, we need to set it to zero, otherwise we cannot distinguish between a value that was or was not present 
# in the dataset after pivoting the data
df_all_selected = df_all_selected.fillna({'value':0.0}) # set null value in value column to 0.0

# sometimes negative number are reported as a positive number, but the "negating" flag is set to 1 in these cases.
# An example where this happens is "PaymentsForDividens" normally, PaymentsForDividens is a negative number in the CF statements, since it is an outflow
# but often (however not always) it appears as a positive number in the data with the negated Flag set to 1.
# Therefore, PaymentsForDividens values can be positive or negative.
# To make the numbers compareable, we have to recalculate the value if it negating is set to 1.
df_all_selected = df_all_selected.withColumn('value', when(df_all_selected.negating == 1, df_all_selected.value * -1).otherwise(df_all_selected.value))

pivot_attrs = ['value'] # column that contains the value which has to be pivoted

def pivot_statement(all_data_df, statement:str, stmtfilter:str):
    all_stmt_data = all_data_df.where("stmt == '" + statement + "'" + stmtfilter).cache()
    
    shutil.rmtree(all_pivoted_folder + "/" + statement,  ignore_errors=True)
    
    # it is important to also include the qrts,report and line into grupping. 
    # - for instance, the IS is often not just for the last quarter, but also for the numbers of quarters since beginning of the fiscal year
    # - the same tag can appear multiple times in a repor, e.g. the IS has cash at the beginning and at end of the period
    grouped_df = all_stmt_data.groupby(["cik","ticker","adsh","form","period", "filed", "fp", "qtrs"])
    
    for attr in pivot_attrs: 
        # using max() is not the best approach. generally, a tag is only contained once in a report, but there are excptions
        # like the CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalents in the CF statement.
        pivoted_df = grouped_df.pivot("tag").max(attr)

        # repartition(1), so that only one file is created
        pivoted_df.repartition(1).write.parquet(all_pivoted_folder + "/" + statement + "/" + attr) 
        
    all_stmt_data.unpersist()

def pivot_statements(all_data_df, statements,filters):
    start = time.time()
    
    for stmt,stmtfilter in zip(statements,filters): 
        print (stmt, end = "")
        section_start = time.time()
        pivot_statement(all_data_df, statement=stmt,stmtfilter = stmtfilter)
        section_duration = time.time() - section_start
        print (": ", section_duration)

    duration = time.time() - start
    print("duration: ", duration)

## 99_Execution

#statements = ['IS','CF','CP','BS','CI','EQ','UN']

statements = ['IS','CF','BS','UN']

# #in case of the IS statement, we are only interested in the qtrs=4 for a 10Ks and qtrs=1 for a 10Qs
# #in case of the CF statement, the relevant information is in qtrs 4 for the 10K or in the appropriate qrtr for every period Q1..Q4
# #  qrtr 0 seems to contain only cash at carrying value at the end of the period
# # in case of the BS, qtrs == 0

filters = [
        " and ((form == '10-K' and qtrs == '4') or (form == '10-Q' and qtrs == '1'))",
        " and ((form == '10-K' and qtrs == '4') or (form == '10-Q' and ((fp  == 'Q1' and qtrs == '1') or (fp  == 'Q2' and qtrs == '2') or (fp  == 'Q3' and qtrs == '3') or (fp  == 'Q4' and qtrs == '4'))))",
        " and qtrs=='0' ",
        " ",
]

pivot_statements(df_all_selected, statements, filters)

spark.stop()

## XX_Trials

just_selected = df_all_selected.where("adsh = '0001558370-20-011113'").cache()
just_selected.show(50, False)

just_is_sel = just_selected.where("stmt == 'IS'")

grouped_df = just_is_sel.groupby(["cik","ticker","adsh","form","period","fp", "qtrs"])

pivoted_df = grouped_df.pivot("tag").max('value')

pd = pivoted_df.toPandas()

pd
