==> 04_02_Uniform_CF.ipynb <==
# default_exp uniformcf

#hide
from nbdev.showdoc import *

#hide
# stellt sicher, dass beim verändern der core library diese wieder neu geladen wird
%load_ext autoreload
%autoreload 2

# Uniform CF

## Basic Settings

# imports
from bfh_mt_hs2020_sec_data.core import * 
from pathlib import Path
from typing import List, Tuple, Union, Set
from pyspark.sql.dataframe import DataFrame
from pyspark.sql.functions import col

import pandas as pd

import shutil          # provides high level file operations
import time            # used to measure execution time
import os
import sys

all_pivot_selected_folder  = "D:/data_mt/03_1_selected_pivot_cols_parquet" 
all_pivoted_folder         = "D:/data_mt/03_2_split_pivot_parquet"         
all_processed_folder       = "D:/data_mt/04_unified/"

col_list =    ["stmt","cik","ticker", "adsh","period","filed", "form","tag","value","report", "line", "fp", "uom"]
pivot_group = ["cik","ticker","adsh","form","period","filed","fp", "qtrs"]

# init Spark
spark = get_spark_session() # Session anlegen
spark # display the moste important information of the session

## 00_Raw_data

# loading the complete unpivoted dataset - if it is needed for debbuging
df_all_selected = spark.read.parquet(all_pivot_selected_folder).cache()

# it happens sometimes, that the data could not be associated with a right sheet (bs, is, cf, ..). in this cases, the data can appea under "UN"
# so if expected information cannot be found in the appropriate statement, we have to look in the un statement
un_pivot_value = load_data(all_pivoted_folder, spark, "UN", "value")
un_pivot_value = un_pivot_value.where("((form == '10-K' and qtrs == '4') or (form == '10-Q' and ((fp  == 'Q1' and qtrs == '1') or (fp  == 'Q2' and qtrs == '2') or (fp  == 'Q3' and qtrs == '3') or (fp  == 'Q4' and qtrs == '4'))))")
un_pivot_pd = un_pivot_value.toPandas()

un_pivot_pd.shape

def prepare_un_values(df_to_merge_into, attr_list):
    # add possible columns from un set to  cf data with prefix cpy_
    attributes = pivot_group[:] # create copy
    attributes.extend(attr_list)

    un_prepared = un_pivot_pd[attributes].copy()
    un_prepared.rename(columns=lambda x: x  if x in pivot_group else ("cpy_" + x), inplace=True)

    return pd.merge(df_to_merge_into, un_prepared, how='left', on=pivot_group)

## 01_CashFlow

Operation
- NetIncomeLoss
- ProfitLoss
- NetCashProvidedByUsedInOperatingActivities: NetIncome + other positions ergibt diese Position

Investing
- NetCashProvidedByUsedInInvestingActivities

Financing activities
- PaymentsForRepurchaseOfCommonStock: Aktienrückkäufe
- PaymentsOfDividends
- NetCashProvidedByUsedInFinancingActivities

Cash Bestand unterschied
- CashAndCashEquivalentsPeriodIncreaseDecrease: increase/decrease in cash

cf_pivot_value = load_data(all_pivoted_folder, spark, "CF", "value")
spark_shape(cf_pivot_value)

cf_pivot_pd = cf_pivot_value.toPandas()

cf_pivot_pd_copy = cf_pivot_pd.copy()

cf_pivot_pd_copy.shape

### Merge Data from "unknown" statement

# merge relevant columns from the UN dataset
cf_pivot_pd_copy = prepare_un_values(cf_pivot_pd_copy, [
    'CashAndCashEquivalentsPeriodIncreaseDecrease',
     # 'CashAndCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect',
    'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect',
    'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseIncludingExchangeRateEffect',   
    # 'CashPeriodIncreaseDecrease',
    # 'CashPeriodIncreaseDecreaseExcludingExchangeRateEffect',
    'NetCashProvidedByUsedInContinuingOperations',
    
    'ProfitLoss',
    'NetIncomeLoss', 
    'NetIncomeLossAvailableToCommonStockholdersBasic',
    'NetCashProvidedByUsedInOperatingActivities',
    'NetCashProvidedByUsedInOperatingActivitiesContinuingOperations',
 
    'NetCashProvidedByUsedInInvestingActivities',
    'NetCashProvidedByUsedInFinancingActivities',
    'NetCashProvidedByUsedInInvestingActivitiesContinuingOperations',
    'CashProvidedByUsedInInvestingActivitiesDiscontinuedOperations',
    
    'NetCashProvidedByUsedInFinancingActivitiesContinuingOperations',
    # 'CashProvidedByUsedInFinancingActivitiesDiscontinuedOperations' ,
])
cf_pivot_pd_copy.shape

### Cash Increase/Decrease
- 'CashAndCashEquivalentsPeriodIncreaseDecrease',
- 'CashAndCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect',
- 'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect',
- 'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseIncludingExchangeRateEffect',
- 'CashPeriodIncreaseDecrease',
- 'CashPeriodIncreaseDecreaseExcludingExchangeRateEffect',
- 'NetCashProvidedByUsedInContinuingOperations'

print_null_count(cf_pivot_pd_copy, ['CashAndCashEquivalentsPeriodIncreaseDecrease'])

cf_pivot_pd_copy['c_CashIncreaseDecrease'] = None

copy_if_not_empty(cf_pivot_pd_copy, 'cpy_CashAndCashEquivalentsPeriodIncreaseDecrease', 'CashAndCashEquivalentsPeriodIncreaseDecrease')
copy_if_not_empty(cf_pivot_pd_copy, 'cpy_CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect', 'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect')
copy_if_not_empty(cf_pivot_pd_copy, 'cpy_CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseIncludingExchangeRateEffect', 'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseIncludingExchangeRateEffect')


copy_if_not_empty(cf_pivot_pd_copy, 'CashAndCashEquivalentsPeriodIncreaseDecrease', 'c_CashIncreaseDecrease')
copy_if_not_empty(cf_pivot_pd_copy, 'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseIncludingExchangeRateEffect', 'c_CashIncreaseDecrease') # either or
copy_if_not_empty(cf_pivot_pd_copy, 'CashCashEquivalentsRestrictedCashAndRestrictedCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect', 'c_CashIncreaseDecrease') # either or
copy_if_not_empty(cf_pivot_pd_copy, 'CashPeriodIncreaseDecrease', 'c_CashIncreaseDecrease') # either or
copy_if_not_empty(cf_pivot_pd_copy, 'CashAndCashEquivalentsPeriodIncreaseDecreaseExcludingExchangeRateEffect', 'c_CashIncreaseDecrease') # either or
copy_if_not_empty(cf_pivot_pd_copy, 'CashPeriodIncreaseDecreaseExcludingExchangeRateEffect', 'c_CashIncreaseDecrease') # either or

print_null_count(cf_pivot_pd_copy, ['c_CashIncreaseDecrease'])

### NetIncomeLoss / Operation CashFlow

print_null_count(cf_pivot_pd_copy, ['NetCashProvidedByUsedInOperatingActivities'])

cf_pivot_pd_copy['c_CashFromOperating'] = None

copy_if_not_empty(cf_pivot_pd_copy, 'cpy_NetCashProvidedByUsedInOperatingActivities', 'NetCashProvidedByUsedInOperatingActivities')
copy_if_not_empty(cf_pivot_pd_copy, 'cpy_NetCashProvidedByUsedInOperatingActivitiesContinuingOperations', 'NetCashProvidedByUsedInOperatingActivitiesContinuingOperations')

copy_if_not_empty(cf_pivot_pd_copy, 'NetCashProvidedByUsedInOperatingActivities', 'c_CashFromOperating')
copy_if_not_empty(cf_pivot_pd_copy, 'NetCashProvidedByUsedInOperatingActivitiesContinuingOperations', 'c_CashFromOperating')

print_null_count(cf_pivot_pd_copy, ['c_CashFromOperating'])

### Investing
- NetCashProvidedByUsedInInvestingActivities

print_null_count(cf_pivot_pd_copy, ['NetCashProvidedByUsedInInvestingActivities'])

cf_pivot_pd_copy['c_CashFromInvesting'] = None

copy_if_not_empty(cf_pivot_pd_copy, 'cpy_NetCashProvidedByUsedInInvestingActivities', 'NetCashProvidedByUsedInInvestingActivities')
copy_if_not_empty(cf_pivot_pd_copy, 'cpy_NetCashProvidedByUsedInInvestingActivitiesContinuingOperations','NetCashProvidedByUsedInInvestingActivitiesContinuingOperations')
copy_if_not_empty(cf_pivot_pd_copy, 'cpy_CashProvidedByUsedInInvestingActivitiesDiscontinuedOperations' ,'CashProvidedByUsedInInvestingActivitiesDiscontinuedOperations')

copy_if_not_empty(cf_pivot_pd_copy, 'NetCashProvidedByUsedInInvestingActivities', 'c_CashFromInvesting')
copy_if_not_empty(cf_pivot_pd_copy, 'NetCashProvidedByUsedInInvestingActivitiesContinuingOperations', 'c_CashFromInvesting')
copy_if_not_empty(cf_pivot_pd_copy, 'CashProvidedByUsedInInvestingActivitiesDiscontinuedOperations', 'c_CashFromInvesting')

set_to_zero_if_null(cf_pivot_pd_copy, 'c_CashFromInvesting') # it is absolutely possible, that there are no investing activities

print_null_count(cf_pivot_pd_copy, ['c_CashFromInvesting'])

### Financing activities

#### NetCashProvidedByUsedInFinancingActivities

print_null_count(cf_pivot_pd_copy, ['NetCashProvidedByUsedInFinancingActivities'])

cf_pivot_pd_copy['c_CashFromFinancing'] = None

copy_if_not_empty(cf_pivot_pd_copy, 'cpy_NetCashProvidedByUsedInFinancingActivities', 'NetCashProvidedByUsedInFinancingActivities')
copy_if_not_empty(cf_pivot_pd_copy, 'cpy_NetCashProvidedByUsedInFinancingActivitiesContinuingOperations','NetCashProvidedByUsedInFinancingActivitiesContinuingOperations')

sum_into_empty_target(cf_pivot_pd_copy, 
                      'NetCashProvidedByUsedInFinancingActivitiesContinuingOperations',
                      'CashProvidedByUsedInFinancingActivitiesDiscontinuedOperations' ,
                      'NetCashProvidedByUsedInFinancingActivities')

copy_if_not_empty(cf_pivot_pd_copy, 'NetCashProvidedByUsedInFinancingActivities', 'c_CashFromFinancing')
copy_if_not_empty(cf_pivot_pd_copy, 'NetCashProvidedByUsedInFinancingActivitiesContinuingOperations', 'c_CashFromFinancing')
copy_if_not_empty(cf_pivot_pd_copy, 'CashProvidedByUsedInFinancingActivitiesDiscontinuedOperations', 'c_CashFromFinancing')

set_to_zero_if_null(cf_pivot_pd_copy, 'c_CashFromFinancing') # it is absolutely possible, that there are no financing activities

print_null_count(cf_pivot_pd_copy, ['c_CashFromFinancing'])

#### PaymentsOfDividends
Simply set to 0.0 if no data is present
 'PaymentsOfDividends',
 'PaymentsOfDividendsCommonStock',
 'PaymentsOfDividendsMinorityInterest',
 'PaymentsOfDividendsPreferredStockAndPreferenceStock',
 'PaymentsOfOrdinaryDividends',

cf_pivot_pd_copy = sum_cols_into_new_target(cf_pivot_pd_copy, 'c_PaymentsOfDividendsTotal', 
                                 ['PaymentsOfDividends',
                                 'PaymentsOfDividendsCommonStock',
                                 'PaymentsOfDividendsMinorityInterest',
                                 'PaymentsOfDividendsPreferredStockAndPreferenceStock',
                                 'PaymentsOfOrdinaryDividends'])

#### PaymentsForRepurchaseOfCommonStock: Stock buybacks
 'PaymentsForRepurchaseOfCommonStock',
 'PaymentsForRepurchaseOfCommonStockForEmployeeTaxWithholdingObligations',
 'PaymentsForRepurchaseOfConvertiblePreferredStock',
 'PaymentsForRepurchaseOfPreferredStockAndPreferenceStock',
 'PaymentsForRepurchaseOfRedeemableConvertiblePreferredStock',
 'PaymentsForRepurchaseOfRedeemablePreferredStock'

cf_pivot_pd_copy = sum_cols_into_new_target(cf_pivot_pd_copy, 'c_PaymentsForRepurchaseOfStockTotal', 
                                 ['PaymentsForRepurchaseOfCommonStock',
                                     'PaymentsForRepurchaseOfCommonStockForEmployeeTaxWithholdingObligations',
                                     'PaymentsForRepurchaseOfConvertiblePreferredStock',
                                     'PaymentsForRepurchaseOfPreferredStockAndPreferenceStock',
                                     'PaymentsForRepurchaseOfRedeemableConvertiblePreferredStock',
                                     'PaymentsForRepurchaseOfRedeemablePreferredStock'])

### Corrections
some Bookings are simply wrong. There are a few cases, when PayedDividends Appear as an inflow, which is wrong. The same holds true for PaymentsForRepurchaseOfStockTotal

cf_pivot_pd_copy.loc[cf_pivot_pd_copy.c_PaymentsOfDividendsTotal>0,'c_PaymentsOfDividendsTotal'] = -cf_pivot_pd_copy.loc[cf_pivot_pd_copy.c_PaymentsOfDividendsTotal>0].c_PaymentsOfDividendsTotal
cf_pivot_pd_copy.loc[cf_pivot_pd_copy.c_PaymentsForRepurchaseOfStockTotal>0,'c_PaymentsForRepurchaseOfStockTotal'] = -cf_pivot_pd_copy.loc[cf_pivot_pd_copy.c_PaymentsForRepurchaseOfStockTotal>0].c_PaymentsForRepurchaseOfStockTotal

### Save

cf_pivot_pd_copy[["cik","ticker", "adsh","period","filed", "form", "qtrs","fp",
                                     'c_CashIncreaseDecrease',
                                     'c_CashFromOperating', 
                                     'c_CashFromInvesting',
                                     'c_CashFromFinancing',
                                     'c_PaymentsOfDividendsTotal',
                                     'c_PaymentsForRepurchaseOfStockTotal']] \
                  .to_csv(all_processed_folder + "04_cf_unified.csv", index=False)

spark.stop()

## XX_Trials

div_data = cf_pivot_pd_copy[cf_pivot_pd_copy.c_PaymentsForRepurchaseOfStockTotal>0][['ticker','cik','adsh', 'c_PaymentsForRepurchaseOfStockTotal',
                                     'PaymentsForRepurchaseOfCommonStock',
                                     'PaymentsForRepurchaseOfCommonStockForEmployeeTaxWithholdingObligations',
                                     'PaymentsForRepurchaseOfConvertiblePreferredStock',
                                     'PaymentsForRepurchaseOfPreferredStockAndPreferenceStock',
                                     'PaymentsForRepurchaseOfRedeemableConvertiblePreferredStock',
                                     'PaymentsForRepurchaseOfRedeemablePreferredStock']]

print(cf_pivot_pd_copy[cf_pivot_pd_copy.c_PaymentsForRepurchaseOfStockTotal>0].shape[0])
print(div_data.shape[0])
print(len(div_data.ticker.unique()))

div_data[div_data.ticker=='AAPL'].sort_values('adsh')

aapl_pd = df_all_selected.where("adsh == '0000320193-17-000070' and stmt = 'CF'").toPandas()

aapl_pd

# cf_pivot_value.select('adsh').distinct().count() # 116482
# cf_pivot_value.select('adsh','form').where('form == "10-K"').distinct().count() # 27556
# cf_pivot_value.select('adsh','form').where('form == "10-Q"').distinct().count() # 88926

# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-K" and qtrs="0"').distinct().count() # 27264
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-K" and qtrs="1"').distinct().count() # 16405
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-K" and qtrs="2"').distinct().count() # 63
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-K" and qtrs="3"').distinct().count() # 44
cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-K" and qtrs="4"').distinct().count() # 27497

# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-Q" and qtrs="0"').distinct().count() # 88076
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-Q" and qtrs="1"').distinct().count() # 86728
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-Q" and qtrs="2"').distinct().count() # 30215
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-Q" and qtrs="3"').distinct().count() # 28150
# cf_pivot_value.select('adsh','form', 'qtrs').where('form == "10-Q" and qtrs="4"').distinct().count() # 185

piv_by_qtrs_pd = cf_pivot_value.select('adsh','fp','form','qtrs').groupby(['adsh','fp','form']).pivot('qtrs').count().toPandas()


#pd_frame = df_all_selected.where("adsh == '0001564590-20-043606' and stmt == 'IS' and qtrs=='1'").toPandas()
#pd_frame = df_all_selected.where("adsh == '0001628279-20-000210'").toPandas()
#pd_frame = df_all_selected.where("adsh == '0001193125-20-213555'").toPandas()
#print(pd_frame.sort_values(['report', 'line']))

pd.set_option('display.max_rows', 20)
piv_by_qtrs_pd[['adsh','fp','form','0','1','2','3','4']]

# 0001392972-17-000128  Q3 / qtrs 0,1,3 -> 3 voll
pd_frame = cf_pivot_value.where("adsh == '0001193125-13-315367'").toPandas()

pd_frame = cf_pivot_value.where("adsh == ' 0001392972-17-000128'").toPandas()
pd.set_option('display.max_rows', pd_frame.shape[0]+1)
pd_frame.sort_values(['qtrs']).dropna(how='all', axis=1)

pd_frame = cf_pivot_value.where("adsh == '0001193125-13-315367'").toPandas()
pd.set_option('display.max_rows', pd_frame.shape[0]+1)
pd_frame.sort_values(['qtrs']).dropna(how='all', axis=1)

pd_frame = cf_pivot_value.where("adsh == '0000946673-19-000012'").toPandas()
pd.set_option('display.max_rows', pd_frame.shape[0]+1)
pd_frame.sort_values(['qtrs']).dropna(how='all', axis=1)

df_all_selected.where("adsh == '0001654954-19-002486' and stmt = 'CF'").orderBy('line').show(50,False)
