==> 09_Final_Preparation.ipynb <==
# default_exp final

#hide
from nbdev.showdoc import *

#hide
# nbdev function - ensures that changed libraries from the project are reloaded
%load_ext autoreload
%autoreload 2

# Final preparations

# imports
from bfh_mt_hs2020_sec_data.core import * 
from pathlib import Path
from typing import List, Tuple, Union, Set
import glob

import pandas as pd
import numpy as np

import shutil          # provides high level file operations
import time            # used to measure execution time
import os
import sys

import matplotlib.pyplot as plt
import seaborn as sns; sns.set()

from yahoo_historical import Fetcher
import yfinance as yf

all_data_local_folder = "./data/"
stock_data_folder     = "D:/data_mt/08_stock_data/"
all_features_folder   = "D:/data_mt/07_features/"

trainingset_folder    = "D:/data_mt/09_training/"

join_group = ["cik","ticker","adsh","period","filed","form","fp"]

## 00_Tools

def load_sec_features():
    return pd.read_csv(all_features_folder + "07_04_features_fp_corrected.csv", header = 0)

def read_additional_info() -> pd.DataFrame:
    return pd.read_csv(stock_data_folder + "08_add_ticker_info_features.csv", sep=',', encoding='utf-8', header=0)

# create a list with all stockdata files
def get_stock_data_files() -> List[str]:
    return glob.glob(stock_data_folder + "**/*_processed.csv",recursive = True)

def get_ticker_from_filename(filename: str) -> str:
    return filename[len(stock_data_folder) + 2 : -14]

## 01_Load data

df_sec = load_sec_features()

## 02_Check Ticker and CIK in financial statements data
it could be that the same Ticker-Symbol is used for several CIKs. This could happen, if a company is delisted or if companies are part of a holding. However, we have to ensure that the mapping between the ticker and cik we are using is unambiguous.

df_ciktik = df_sec[['cik','ticker']]
df_ciktik_unique = df_ciktik.drop_duplicates()

df_cik_to_tickers = df_ciktik_unique['ticker'].value_counts()
# this shows how often a ticker is is used for how many ciks
# what we would want to see is one row starting with "1".  
# If there multiple rows, wie now that there are tickers which seem to be used for different ciks.
df_cik_to_tickers.value_counts() 

used_mulitple_times = df_cik_to_tickers[df_cik_to_tickers > 1].index.to_list()
used_mulitple_times

df_ciktik_unique[df_ciktik_unique.ticker.isin(used_mulitple_times)].sort_values('ticker')

## 03_Check completeness of data
there are now 3 datasets with features
- 07_04_features_fp_corrected.csv contains comparable data from the business reports
- [ticker]_processed.csv contain the historical stockdata
- 08_add_ticker_info_features.csv contains additional information for a company

for the training, we can only use the data of companies that are present in all 3 datasets 

sd_files = get_stock_data_files()
ticker_sd_files = [get_ticker_from_filename(x) for x in sd_files]

set_ticker_add_info = set(add_info.ticker.unique())
set_ticker_df_sec = set(df_sec.ticker.unique())
set_ticker_sd = set(ticker_sd_files)

print(len(set_ticker_add_info))
print(len(set_ticker_df_sec))
print(len(set_ticker_sd))

# define which tickers are present in all 3 datasets
intersected = set_ticker_add_info.intersection(set_ticker_df_sec).intersection(set_ticker_sd)
len(intersected)

### save/copy data of companies present in all 3 datasets to a special folder

add_info_cleaned = add_info[add_info.ticker.isin(intersected)].sort_values('ticker')
add_info_cleaned.to_csv(trainingset_folder + "company_info.csv", header=True, index=False)

df_sec_cleaned = df_sec[df_sec.ticker.isin(intersected)].sort_values(['ticker','period'])
df_sec_cleaned.to_csv(trainingset_folder + "company_reports.csv", header=True, index=False)

import shutil
import string
for char in string.ascii_uppercase:
    folder = trainingset_folder + "stocks/" + char + "/"
    directory = os.path.dirname(folder)
    if not os.path.exists(directory):
        os.makedirs(directory)
        
for file in sd_files:
    ticker = get_ticker_from_filename(file)
    if ticker in intersected:
        shutil.copy(file, trainingset_folder + "stocks/" + ticker[0] + "/" + ticker +".csv")

## 04_Combine Stockdata into a single dataset
In order to be able to normalize the stockdata, we need to have them all in a single dataset. <br>
Since we don't know how we will use them, we store the data with spark as a parquet dataset and a pure csv dataset.

# init Spark
spark = get_spark_session() # Session anlegen
spark # display the moste important information of the session

start = time.time()
df_all = spark.read.csv(trainingset_folder + "stocks/*/*", header=True, dateFormat="yyyy-MM-dd")

from pyspark.sql.functions import col
df_all_partioned = df_all.repartition(16, col("ticker"))
df_all_partioned.write.parquet(trainingset_folder + "/all_stock_data")

pd_df = df_all.toPandas()

pd_df.sort_values(['ticker','Date']).to_csv(trainingset_folder + "all_stock_data.csv", header=True, index=False)

spark.stop()

## XX_Trials

print(len(add_info.sector.unique()))
print(len(add_info.industry.unique()))

add_info.sector.value_counts()

industries = add_info.industry.value_counts()
industries[industries > 20].index.tolist()

add_info[add_info.marketCap_rank < 15].sort_values('marketCap_rank')

industries = add_info.industry.unique()

reit_list = list(filter(lambda k: 'REIT' in k, industries))
real_list = list(filter(lambda k: 'RealEstate' in k, industries))
gasoil_list = list(filter(lambda k: 'Gas' in k, industries))
health_list = list(filter(lambda k: 'Health' in k, industries))
insurance_list = list(filter(lambda k: 'Insurance' in k, industries))
rest = set(industries) - set(reit_list) - set(gasoil_list) - set(real_list) - set(health_list) - set(insurance_list)

print(len(reit_list))
print(len(gasoil_list))
print(len(real_list))
print(len(health_list))
print(len(insurance_list))
print(len(rest))
print(rest)

adf = read_additional_info()
adf[adf.message.isna()]
